Fixed opt_fuzz in cgi and graph cmds from int to double; fixed travis clang build

+++ duc/cmd-cgi.c
@@ -25,7 +25,7 @@ struct param {
 
 static char *opt_database = NULL;
 static int opt_size = 800;
-static int opt_fuzz = 0.7;
+static double opt_fuzz = 0.7;
 static int opt_levels = 4;
 static char *opt_output = NULL;
 static char *opt_palette = NULL;
